move group shape util into shapes folder
