+++ packages/editor/src/lib/editor/shapes/ArrowShape/ArrowShapeUtil/arrow/curved-arrow.ts
@@ -13,13 +13,13 @@ import {
 	Vec2d,
 	VecLike,
 } from '@tldraw/primitives'
-import { TLArrowShape } from '@tldraw/tlschema'
 import {
 	BOUND_ARROW_OFFSET,
 	MIN_ARROW_LENGTH,
 	WAY_TOO_BIG_ARROW_BEND_FACTOR,
 } from '../../../../../constants'
 import { Editor } from '../../../../Editor'
+import { TLArrowShape } from '../../arrowShapeTypes'
 import { ArcInfo, ArrowInfo } from './arrow-types'
 import { getArrowTerminalsInArrowSpace, getBoundShapeInfoForTerminal } from './shared'
 import { getStraightArrowInfo } from './straight-arrow'
