+++ packages/editor/src/lib/editor/shapes/ArrowShape/arrowShape.ts
@@ -5,6 +5,7 @@ import { ArrowShapeUtil } from './ArrowShapeUtil/ArrowShapeUtil'
 import { arrowShapeMigrations } from './arrowShapeMigrations'
 import { arrowShapeValidator } from './arrowShapeValidator'
 
+/** @public */
 export const arrowShape: TLShapeInfo & { tool?: TLStateNodeConstructor } = {
 	util: ArrowShapeUtil,
 	tool: ArrowShapeTool,
