+++ packages/editor/src/lib/editor/shapes/ArrowShape/arrowShapeTypes.ts
@@ -1,31 +1,13 @@
 import {
-	SetValue,
+	TLArrowTerminal,
 	TLArrowheadType,
 	TLBaseShape,
 	TLColorType,
 	TLDashType,
 	TLFillType,
 	TLFontType,
-	TLShapeId,
 	TLSizeType,
-	Vec2dModel,
 } from '@tldraw/tlschema'
-
-/** @public */
-export const TL_ARROW_TERMINAL_TYPE = new Set(['binding', 'point'] as const)
-
-/** @public */
-export type TLArrowTerminalType = SetValue<typeof TL_ARROW_TERMINAL_TYPE>
-
-/** @public */
-export type TLArrowTerminal =
-	| {
-			type: 'binding'
-			boundShapeId: TLShapeId
-			normalizedAnchor: Vec2dModel
-			isExact: boolean
-	  }
-	| { type: 'point'; x: number; y: number }
 
 /** @public */
 export type TLArrowShapeProps = {
