+++ packages/editor/src/lib/editor/shapes/HighlightShape/HighlightShapeTool/HighlightShapeTool.test.ts
