+++ packages/editor/src/lib/editor/shapes/HighlightShape/highlightShapeMigrations.ts
@@ -1,0 +1,4 @@
+import { defineMigrations } from '@tldraw/store'
+
+/** @internal */
+export const highlightShapeMigrations = defineMigrations({})
