+++ packages/editor/src/lib/editor/shapes/HighlightShape/highlightShapeTypes.ts
@@ -1,0 +1,14 @@
+import { TLBaseShape, TLColorType, TLSizeType } from '@tldraw/tlschema'
+import { TLDrawShapeSegment } from '../DrawShape/drawShapeTypes'
+
+/** @public */
+export type TLHighlightShapeProps = {
+	color: TLColorType
+	size: TLSizeType
+	segments: TLDrawShapeSegment[]
+	isComplete: boolean
+	isPen: boolean
+}
+
+/** @public */
+export type TLHighlightShape = TLBaseShape<'highlight', TLHighlightShapeProps>
