+++ packages/tlschema/src/misc/TLArrowTerminal.ts
@@ -1,0 +1,36 @@
+import { T } from '@tldraw/validate'
+import { TLShapeId } from '../records/TLShape'
+import { shapeIdValidator } from '../shapes/TLBaseShape'
+import { SetValue } from '../util-types'
+import { Vec2dModel } from './geometry-types'
+
+/** @public */
+export const TL_ARROW_TERMINAL_TYPE = new Set(['binding', 'point'] as const)
+
+/** @public */
+export type TLArrowTerminalType = SetValue<typeof TL_ARROW_TERMINAL_TYPE>
+
+/** @public */
+export type TLArrowTerminal =
+	| {
+			type: 'binding'
+			boundShapeId: TLShapeId
+			normalizedAnchor: Vec2dModel
+			isExact: boolean
+	  }
+	| { type: 'point'; x: number; y: number }
+
+/** @public */
+export const arrowTerminalValidator: T.Validator<TLArrowTerminal> = T.union('type', {
+	binding: T.object({
+		type: T.literal('binding'),
+		boundShapeId: shapeIdValidator,
+		normalizedAnchor: T.point,
+		isExact: T.boolean,
+	}),
+	point: T.object({
+		type: T.literal('point'),
+		x: T.number,
+		y: T.number,
+	}),
+})
