+++ packages/tlschema/src/records/TLShape.ts
@@ -1,14 +1,11 @@
 import { defineMigrations, RecordId, UnknownRecord } from '@tldraw/store'
 import { nanoid } from 'nanoid'
-import { TLArrowShape } from '../shapes/TLArrowShape'
 import { TLBaseShape } from '../shapes/TLBaseShape'
 import { TLBookmarkShape } from '../shapes/TLBookmarkShape'
-import { TLDrawShape } from '../shapes/TLDrawShape'
 import { TLEmbedShape } from '../shapes/TLEmbedShape'
 import { TLFrameShape } from '../shapes/TLFrameShape'
 import { TLGeoShape } from '../shapes/TLGeoShape'
 import { TLGroupShape } from '../shapes/TLGroupShape'
-import { TLHighlightShape } from '../shapes/TLHighlightShape'
 import { TLIconShape } from '../shapes/TLIconShape'
 import { TLImageShape } from '../shapes/TLImageShape'
 import { TLLineShape } from '../shapes/TLLineShape'
@@ -22,9 +19,7 @@ import { TLPageId } from './TLPage'
  *
  * @public */
 export type TLDefaultShape =
-	| TLArrowShape
 	| TLBookmarkShape
-	| TLDrawShape
 	| TLEmbedShape
 	| TLFrameShape
 	| TLGeoShape
@@ -35,7 +30,6 @@ export type TLDefaultShape =
 	| TLTextShape
 	| TLVideoShape
 	| TLIconShape
-	| TLHighlightShape
 
 /**
  * A type for a shape that is available in the editor but whose type is
