+++ packages/tlschema/src/styles/style-types.ts
@@ -4,28 +4,28 @@ import {
 	TLDashStyle,
 	TLFillStyle,
 	TLFontStyle,
+	TLShapeProps,
 	TLSizeStyle,
-	TLStyleType,
 } from '..'
-import { TLShapeProps } from '../records/TLShape'
 import { TLArrowheadEndStyle, TLArrowheadStartStyle } from './TLArrowheadStyle'
+import { TLBaseStyle, TLStyleType } from './TLBaseStyle'
 import { TLGeoStyle } from './TLGeoStyle'
 import { TLSplineStyle } from './TLSplineStyle'
 import { TLVerticalAlignStyle } from './TLVerticalAlignStyle'
 
 /** @public */
-export type TLStyleItem =
-	| TLColorStyle
-	| TLFillStyle
-	| TLDashStyle
-	| TLSizeStyle
-	| TLFontStyle
-	| TLAlignStyle
-	| TLVerticalAlignStyle
-	| TLGeoStyle
-	| TLArrowheadStartStyle
-	| TLArrowheadEndStyle
-	| TLSplineStyle
+export type TLStyleItem = TLBaseStyle
+// | TLColorStyle
+// | TLFillStyle
+// | TLDashStyle
+// | TLSizeStyle
+// | TLFontStyle
+// | TLAlignStyle
+// | TLVerticalAlignStyle
+// | TLGeoStyle
+// | TLArrowheadStartStyle
+// | TLArrowheadEndStyle
+// | TLSplineStyle
 //	| TLIconStyle
 
 /** @public */
@@ -44,4 +44,4 @@ export interface TLStyleCollections {
 }
 
 /** @public */
-export type TLStyleProps = Pick<TLShapeProps, TLStyleType>
+export type TLStyleProps = Pick<TLShapeProps, TLStyleType> // @TODO
